Separate tests & side-effects
